backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.0
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.115.2
h11==0.14.0
idna==3.10
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
sniffio==1.3.1
starlette==0.40.0
typing_extensions==4.12.2
uvicorn==0.32.0
